+++ src/Scene/lighting/PointLight.java
@@ -3,7 +3,7 @@ package Scene.lighting;
 import java.awt.*;
 
 public class PointLight extends Light {
-    public PointLight(double x, double y, double z, Color color) {
+    public PointLight(float x, float y, float z, Color color) {
         super(x, y, z, color);
     }
 }
