+++ src/Scene/objects/GLTF.java
@@ -5,7 +5,7 @@ import org.json.simple.JSONObject;
 import java.awt.*;
 
 public class GLTF extends Shape {
-    public GLTF(double x, double y, double z, JSONObject size, Color colour) {
-        super(x, y, z, (double) size.get("scale"), colour);
+    public GLTF(float x, float y, float z, JSONObject size, Color colour) {
+        super(x, y, z, (float) size.get("scale"), colour);
     }
 }
